ntdll, server: Revert to old implementation of hung queue detection.
By manually notifying the server every time we enter and exit a message wait.

The hung queue logic keeps breaking. In the case of bug zfigura/wine#9 it was
breaking because we were waiting for more than 5 seconds on our queue and
then someone sent us a message with SMTO_ABORTIFHUNG. Just stop fighting
against the server and try to coöperate with it instead. It takes two extra
server calls, but ideally the GUI thread isn't going to be in the same sort
of performance-critical code that this patchset was written for.